Show detailed documentations of scMultiSim's parameters

Show detailed documentations of scMultiSim's parameters

Usage

scmultisim_help(topic = NULL)

Arguments

topic

Can be options, dynamic.GRN, or cci

Value

none

Examples

scmultisim_help()
#> Call scmultisim_help(topic) where topic can be "options" or an option na
#> 	me. Printing help for options by default.
#> scMultiSim  v1.0.0
#> 
#> [GENERAL]
#> 
#> rand.seed  (default: 0)
#> 	scMultiSim should produce the same result if all other parameters are th
#> 	e same.
#> 	The value should be a numeric.
#> threads  (default: 1)
#> 	Set to larger than 1 to use multithreading for some part of the simulati
#> 	on.
#> 	The value should be an integer between 1 and 4096.
#> speed.up  (default: FALSE)
#> 	Use experimental speed and memory optimization.
#> 	The value should be a logical.
#> 
#> [GENE]
#> 
#> GRN  (default: NULL)
#> 	The GRN network.
#> 	It should be a data frame with 3 columns (target, regulator, effect). Su
#> 	pply NA to disable the GRN effect.
#> grn.effect  (default: 1)
#> 	Overall strength of the GRN effect on the expression. Different from the
#> 	 effect column in the GRN data frame, which is the relative effect of ea
#> 	ch TF-target pair.
#> 	The value should be a numeric between 0 and Inf.
#> num.genes  (default: NULL)
#> 	Number of genes if GRN is disabled.
#> 	The value should be an integer between 1 and Inf.
#> unregulated.gene.ratio  (default: 0.1)
#> 	Ratio of unreulated to regulated genes. Extra unregulated genes will be 
#> 	simulated in addition to the genes in GRN.
#> 	The value should be a numeric between 0 and 1.
#> giv.mean  (default: 0)
#> 	Mean of the Gene Identity Vectors.
#> 	The value should be a numeric between -Inf and Inf.
#> giv.prob  (default: 0.3)
#> 	Probability of non-zero values in the Gene Identity Vectors.
#> 	The value should be a numeric between 0 and 1.
#> giv.sd  (default: 1)
#> 	Stddev of the Gene Identity Vectors.
#> 	The value should be a numeric between 0 and Inf.
#> hge.range  (default: 1)
#> 	Only choose highly expressed genes after this range.
#> 	The value should be a numeric between 1 and Inf.
#> hge.prop  (default: 0)
#> 	Propotion of highly expressed genes.
#> 	The value should be a numeric between 0 and 1.
#> hge.mean  (default: 5)
#> 	Scale of highly expressed genes.
#> 	The value should be a numeric between 1 and Inf.
#> hge.sd  (default: 1)
#> 	Variation of highly expressed genes.
#> 	The value should be a numeric between 0 and Inf.
#> hge.max.var  (default: 500)
#> 	Genes with higher variation will not be selected as highly expressed gen
#> 	es.
#> 	The value should be a numeric between 0 and Inf.
#> dynamic.GRN  (default: NA)
#> 	Specification of the dynamic GRN. See scmultisim_help("dynamic.GRN") for
#> 	 details.
#> 
#> [CELL]
#> 
#> num.cells  (default: 1000)
#> 	Total number of cells from all populations.
#> 	The value should be an integer between 0 and Inf.
#> tree  (default: Phyla5())
#> 	A tree defining relationship between populations.
#> discrete.cif  (default: FALSE)
#> 	Whether the cell population is discrete.
#> 	The value should be a logical.
#> discrete.pop.size  (default: NA)
#> 	Specify the cell numbers in each population.
#> 	the value should be an integer vector
#> discrete.min.pop.size  (default: 70)
#> 	Size of the smallest discrete cell population.
#> 	The value should be a numeric.
#> discrete.min.pop.index  (default: 1)
#> 	Index of the smallest discrete cell population.
#> 	The value should be an integer between 0 and Inf.
#> num.cifs  (default: 50)
#> 	Number of Cell Identity Factors for each kinetic parameter.
#> 	The value should be a numeric.
#> diff.cif.fraction  (default: 0.9)
#> 	Fraction of CIFs which are differential factors between cell types.
#> 	The value should be a numeric between 0 and 1.
#> cif.center  (default: 1)
#> 	Mean of the CIF values.
#> 	The value should be a numeric.
#> cif.sigma  (default: 0.1)
#> 	Stddev of the CIF values.
#> 	The value should be a numeric between 0 and Inf.
#> use.impulse  (default: FALSE)
#> 	Use the impulse model when generating the continuous CIF.
#> 	The value should be a logical.
#> 
#> [SIMULATION - ATAC]
#> 
#> atac.effect  (default: 0.5)
#> 	The influence of chromatin accessability data on gene expression.
#> 	The value should be a numeric between 0 and 1.
#> region.distrib  (default: c(0.1, 0.5, 0.4))
#> 	The probability that a gene is regulated by respectively 0, 1, 2 consecu
#> 	tive regions.
#> 	the value should be a vector with 3 elements sum to 1
#> atac.p_zero  (default: 0.8)
#> 	The proportion of 0s we see in the ATAC-seq data.
#> atac.density  (default: NA)
#> 	Density of the non-zero ATAC-seq values. Use atac_dens_nonzero() to gene
#> 	rate.
#> 	the value should be a density object.
#> riv.mean  (default: 0)
#> 	Mean of the Region Identity Vectors.
#> 	The value should be a numeric between 0 and Inf.
#> riv.prob  (default: 0.3)
#> 	Probability of non-zero values in the Region Identity Vectors.
#> 	The value should be a numeric between 0 and 1.
#> riv.sd  (default: 1)
#> 	Stddev of the Region Identity Vectors.
#> 	The value should be a numeric between 0 and Inf.
#> 
#> [SIMULATION - RNA]
#> 
#> vary  (default: s)
#> 	Which kinetic parameters have differential CIFs.
#> 	The value should be one of [all, kon, koff, s, except_kon, except_koff, 
#> 	except_s].
#> bimod  (default: 0)
#> 	Adjust the bimodality of gene expression, thus controlling intrinsic var
#> 	iation.
#> 	The value should be a numeric between 0 and 1.
#> scale.s  (default: 1)
#> 	Scale of the s parameter. Use smaller value for cell types known to be s
#> 	mall (like naive cells). When discrete.cif = T, it can be a vector speci
#> 	fying the scale.s for each cluster.
#> intrinsic.noise  (default: 1)
#> 	The weight assigned to the random sample from the Beta-Poisson distribut
#> 	ion, where the weight of the Beta-Poisson mean value is given a weight o
#> 	f (1 - intrinsic.noise).
#> 	The value should be a numeric between 0 and 1.
#> 
#> [SIMULATION - KINETIC MODEL]
#> 
#> do.velocity  (default: FALSE)
#> 	Simulate using the whole kinetic model and generate RNA velocity data.
#> 	The value should be a logical.
#> beta  (default: 0.4)
#> 	Splicing rate of each gene in the kinetic model.
#> 	The value should be a numeric.
#> d  (default: 1)
#> 	Degradation rate of each gene in the kinetic model.
#> 	The value should be a numeric.
#> num.cycles  (default: 3)
#> 	For velocity mode, the number of cycles run before sampling the gene exp
#> 	ression of a cell.
#> 	The value should be an integer between 1 and Inf.
#> cycle.len  (default: 1)
#> 	For velocity mode, a factor multiplied by the expected time to transitio
#> 	n from kon to koff and back to form the the length of a cycle.
#> 	The value should be a numeric between 0 and Inf.
#> mod.cif.giv  (default: NA)
#> 	Modify the generated CIF and GIV. The function takes four arguments: the
#> 	 kinetic parameter index (1=kon, 2=koff, 3=s), the current CIF matrix, t
#> 	he GIV matrix, and the cell metadata dataframe. It should return a list 
#> 	of two elements: the modified CIF matrix and the modified GIV matrix.
#> 	should be a function
#> ext.cif.giv  (default: NA)
#> 	Add customized CIF and GIV. The function takes one argument, the kinetic
#> 	 parameter index (1=kon, 2=koff, 3=s). It should return a list of two el
#> 	ements: the extra CIF matrix (n_extra_cif x n_cells) and the GIV matrix 
#> 	(n_genes x n_extra_cif). Return NULL for no extra CIF and GIV.
#> 	should be a function
#> 
#> [SIMULATION - SPATIAL]
#> 
#> cci  (default: NA)
#> 	The regulation network for spatial cell-cell interaction.
#> 	Enables cell-cell interaction. See scmultisim_help("cci") for details.
#> NULL